requirements.txt
apache-airflow==1.10.9
kafka-python==2.0.2
pyspark==3.1.1
